Refactor Blog components to improve type safety and enhance PostMetadata rendering
- Updated Blog component to use the correct path for blog post links, ensuring consistency with the new PATHS constant.
- Changed BlogPost type import to a more appropriate location for better organization.
- Enhanced PostMetadata component to conditionally render author and date, improving readability and maintainability with clsx for dynamic class management.

This refactor aims to streamline the code structure and improve the overall developer experience while maintaining functionality.
